Buffer geometry is now used
Additionally, to improve culling, instead of always filling in the sides of chunks, an extra row of voxels is included in each dimension of the chunk, which overlap with the neighboring chunk (making the data 17^3 instead of 16^3). This is a little weird but can save up to 2000 tris per chunk. Wowza!